move test to store-level instead of group-level
